Updated OpenVPN server configuraion scripts.

+++ make-ca.sh
@@ -6,7 +6,15 @@
 # Create a directory for CA files
 if [ -d $CA_ROOT ]
 then
-    rm -rf $CA_ROOT
+    read -p "The CA already exists, do you want to recreate it? " ans
+    
+    if echo $ans | grep -q '^[yY]'
+    then 
+        echo "Removing the CA database at \"$CA_ROOT\"."
+        rm -rf $CA_ROOT
+    else
+        exit 0
+    fi
 fi
 
 mkdir $CA_ROOT
@@ -15,7 +23,7 @@ mkdir $CA_ROOT
 touch "$CA_ROOT/index.txt" "$CA_ROOT/index.txt.attr"
 echo 01 > "$CA_ROOT/serial"
 mkdir -m 755 "$CA_ROOT/newcerts" "$CA_ROOT/certs"
-mkdir -m 750 "$CA_ROOT/private" "$CA_ROOT/configs"
+mkdir -m 750 "$CA_ROOT/private" "$CA_ROOT/profiles"
 
 # Write .rnd file (if supported)
 openssl rand -writerand $CA_ROOT/.rnd 2>&1 1>&-
@@ -24,7 +32,14 @@ openssl rand -writerand $CA_ROOT/.rnd 2>&1 1>&-
 openssl req -x509 -days 3650 -out "$CA_CERT" -newkey rsa:2048 -nodes -keyout "$CA_KEY" -config ca.conf -extensions v3_ca
 
 # Generate DH parameter file.
-openssl dhparam -out "$CA_ROOT/dh.pem" 2048
+if [ -f "$OPENVPN_BASEDIR/dh.pem" ]
+then
+    read -p "The Diffie-Helman parameter file already exists, would you like to recreate it? " ans
+    if echo $ans | grep -q '^[Yy]'
+    then
+        sudo openssl dhparam -out "$OPENVPN_BASEDIR/dh.pem" 2048
+    fi
+fi
 
 # Generate static TLS key
 openvpn --genkey --secret "$CA_ROOT/ta.key"
@@ -43,7 +58,7 @@ SUBJ_CN="$SERVER_NAME"
 openssl req -out "$REQ_FILE" -newkey rsa:2048 -nodes -keyout "$KEY_FILE" -config ca.conf
 
 # Sign the request
-openssl ca -in "$REQ_FILE" -out "$CERT_FILE" -config ca.conf -extensions server_ext -batch
+openssl ca -in "$REQ_FILE" -out "$CERT_FILE" -notext -config ca.conf -extensions server_ext -batch
 
 # Remove the request file
 rm -f "$REQ_FILE"
@@ -56,9 +71,4 @@ then
     sudo cp $KEY_FILE $OPENVPN_BASEDIR/server.key
     sudo cp $CA_CRL $OPENVPN_BASEDIR/crl.pem
     sudo cp $CA_ROOT/ta.key $OPENVPN_BASEDIR
-    
-    if [ ! -f $OPENVPN_BASEDIR/dh.pem ]
-    then
-        sudo cp $CA_ROOT/dh.pem $OPENVPN_BASEDIR
-    fi
 fi
